requirements.txt
youtube_dl
isodate
bs4
psutil
lxml
parsedatetime
python-dateutil
timeago
pytz
discord.py
ujson
asyncpg
sentry-sdk
aiocontextvars
uvloop